extracted interface
