+++ frontend/src/interfaces/SignupFormData.ts
@@ -1,0 +1,8 @@
+interface SignupFormData {
+  username: string;
+  email: string;
+  password: string;
+  confirmPassword: string;
+}
+
+export default SignupFormData;
